+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Types of changes are:
 6. Security - For security features
 
 ## Unreleased
+
+## 1.0.26 - 2025-03-12
 ### Added
 - Created `Modal` component
 - Created `Drawer` component